requirements.txt
numpy~=1.23.2
pandas==2.0.3
Pillow~=10.2.0
opencv-python==4.9.0.80
tensorflow==2.9.0
keras==2.9.0
python-dotenv~=1.0.1
deepface~=0.0.85
fastapi~=0.110.0
sqlmodel~=0.0.16
matplotlib~=3.7.5
pymysql
cryptography
pydantic~=2.6.2
typing_extensions~=4.9.0
uvicorn~=0.27.1
sqlalchemy~=2.0.28
starlette~=0.36.3